refactor: usePreviewImage 공통 훅으로 분리 #185
