+++ frontend/src/common/hooks/usePreviewImage.ts
@@ -1,0 +1,20 @@
+import { useState } from 'react';
+
+const usePreviewImage = () => {
+  const [previewUrl, setPreviewUrl] = useState<string | null>(null);
+
+  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const file = e.target.files?.[0];
+
+    if (!file) {
+      return;
+    }
+
+    const fileUrl = URL.createObjectURL(file);
+    setPreviewUrl(fileUrl);
+  };
+
+  return { previewUrl, handleImageChange };
+};
+
+export default usePreviewImage;
